[wip] Support snappy compression for wal archive/restore
It works but... it feels like a compression abstraction would be
useful at this point so that the code doesn't have to care about
the details and we can just add compression implementations in one
place (bearing in mind there are other compression algorithms we'd
like to add too).

We'll keep adding stuff for now and see if there's an obvious
refactor down the line.