Add Barman Cloud option for specifying Azure credential
Adds the `--credential` option to the Barman Cloud scripts. The option
can be set to the following values:

* `azure-cli`: AzureCliCredential, which uses the token obtained by
  logging into azure via the CLI using `az login`.
* `managed-identity`: ManagedIdentityCredential, which uses the Azure
  Managed Identity endpoint to fetch a token (Barman Cloud must be
  running on an Azure resource which supports managed identities for this
  to work).

This improves the user experience when authenticating using either
Azure CLI tokens or managed identities, both of which result in
excessive log noise and latency when handled by `DefaultAzureCredential`.

If the `--credential` option is not set then the existing behaviour is
applied - the environment is checked for the connection string, access
key and SAS token values and if none are found it attempts to use
`DefaultAzureCredential`.