[wip] DRY DecompressingIO
The implementation of decompressing streaming IO is the same for AWS and
Azure aside from the details of how they actually read the compressed
bytes from the stream.

We therefore move them into a mixin which can be inherited from when
subclassing the Streaming IO response for each cloud provider. The
provider-specific logic to read the compressed bytes from the stream is
then added in the subclass.